+++ src/KsWare.AppVeyorClient/UI/PanelProjectSelector/ProjectSelectorVM.cs
@@ -6,6 +6,7 @@ using KsWare.AppVeyorClient.UI.App;
 using KsWare.AppVeyorClient.UI.ViewModels;
 using KsWare.Presentation;
 using KsWare.Presentation.ViewModelFramework;
+using KsWare.Presentation.ViewModelFramework.Providers;
 
 namespace KsWare.AppVeyorClient.UI.PanelProjectSelector {
 
@@ -28,9 +29,12 @@ namespace KsWare.AppVeyorClient.UI.PanelProjectSelector {
 				if (!Client.Base.HasToken) return;
 				var projects=await AppVM.Client.Project.GetProjects();
 				Projects.MːData=projects;
+				((ErrorProvider)Metadata.ErrorProvider).ResetError();
 			}
 			catch (Exception ex) {
+				((ErrorProvider) Metadata.ErrorProvider).SetError($"Could not receive projects.\n{ex.Message}");
 				Debug.WriteLine($"Initialization error on {GetType().FullName}.\n{ex}");
+				//Metadata.ErrorProvider
 				if(Debugger.IsAttached) Debugger.Break();
 			}
 		}