+++ ChangeLog.md
@@ -2,6 +2,7 @@
 - Context menu for YAML editor
 - Syntax highlighting for YAML editor, PowerShell editor
 - Folding for YAML editor
+- reference KsWare.Presentation as MVVM Framework
 # 0.1
 - AppVeyor API
 - YAML configuration Editor